+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/KeyAttestationRequired.kt
@@ -2,6 +2,8 @@ package at.asitplus.openid
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.days
 
 @Serializable
 data class KeyAttestationRequired(
@@ -16,4 +18,13 @@ data class KeyAttestationRequired(
      */
     @SerialName("user_authentication")
     val userAuthentication: Collection<String>? = null,
+
+    /**
+     * OPTIONAL. A duration specifying a PID or Attestation Provider's preference for the remaining validity period of the WUA it receives during issuance, in seconds.
+     * A Wallet Provider SHALL ensure that a Wallet Unit can always present a WUA with a remaining validity period of at least 31 days for their WSCD.
+     * See: https://github.com/eu-digital-identity-wallet/eudi-doc-standards-and-technical-specifications/blob/main/docs/technical-specifications/ts3-wallet-unit-attestation.md
+     */
+    @SerialName("preferred_ttl")
+    @Serializable(with = DurationSecondsIntSerializer::class)
+    val preferredTtl: Duration? = 31.days
 )