+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OAuth2KtorClientTest.kt
@@ -24,6 +24,7 @@ import at.asitplus.wallet.lib.oauth2.OAuth2Client
 import at.asitplus.wallet.lib.oauth2.SimpleAuthorizationService
 import at.asitplus.wallet.lib.oauth2.TokenService
 import at.asitplus.wallet.lib.oidvci.BuildClientAttestationJwt
+import at.asitplus.wallet.lib.oidvci.BuildClientAttestationPoPJwt
 import at.asitplus.wallet.lib.oidvci.CredentialAuthorizationServiceStrategy
 import at.asitplus.wallet.lib.oidvci.decodeFromPostBody
 import at.asitplus.wallet.lib.oidvci.decodeFromUrlQuery
@@ -36,6 +37,7 @@ import io.ktor.client.engine.mock.*
 import io.ktor.client.request.*
 import io.ktor.http.*
 import io.ktor.util.*
+import kotlin.time.Duration.Companion.minutes
 
 val OAuth2KtorClientTest by testSuite {
 
@@ -129,15 +131,26 @@ val OAuth2KtorClientTest by testSuite {
             authorizationService = authorizationService,
             client = OAuth2KtorClient(
                 engine = mockEngine,
-                loadClientAttestationJwt = {
-                    BuildClientAttestationJwt(
-                        SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
-                        clientId = clientId,
-                        issuer = "issuer",
-                        clientKey = clientAuthKeyMaterial.jsonWebKey
-                    ).serialize()
+                loadInstanceAttestation = {
+                    catching {
+                        BuildClientAttestationJwt(
+                            SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
+                            clientId = clientId,
+                            issuer = "issuer",
+                            clientKey = clientAuthKeyMaterial.jsonWebKey
+                        )
+                    }
                 },
-                signClientAttestationPop = SignJwt(clientAuthKeyMaterial, JwsHeaderNone()),
+                loadInstanceAttestationPop = {
+                    catching {
+                        BuildClientAttestationPoPJwt(
+                            SignJwt(clientAuthKeyMaterial, JwsHeaderNone()),
+                            clientId = clientId,
+                            audience = publicContext,
+                            lifetime = 10.minutes,
+                        )
+                    }
+                },
                 signDpop = SignJwt(dpopKeyMaterial, JwsHeaderCertOrJwk()),
                 dpopAlgorithm = dpopKeyMaterial.signatureAlgorithm.toJwsAlgorithm().getOrThrow(),
                 oAuth2Client = OAuth2Client(clientId = clientId),