+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientExternalAuthorizationServerTest.kt
@@ -10,6 +10,7 @@ import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.RequestParameters
 import at.asitplus.openid.TokenIntrospectionRequest
 import at.asitplus.openid.TokenRequestParameters
+import at.asitplus.signum.indispensable.josef.JsonWebToken
 import at.asitplus.signum.indispensable.josef.toJwsAlgorithm
 import at.asitplus.wallet.eupid.EuPidScheme
 import at.asitplus.wallet.lib.agent.ClaimToBeIssued
@@ -40,6 +41,7 @@ import at.asitplus.wallet.lib.oauth2.OAuth2Client
 import at.asitplus.wallet.lib.oauth2.SimpleAuthorizationService
 import at.asitplus.wallet.lib.oauth2.TokenService
 import at.asitplus.wallet.lib.oidvci.BuildClientAttestationJwt
+import at.asitplus.wallet.lib.oidvci.BuildClientAttestationPoPJwt
 import at.asitplus.wallet.lib.oidvci.CredentialAuthorizationServiceStrategy
 import at.asitplus.wallet.lib.oidvci.CredentialDataProviderFun
 import at.asitplus.wallet.lib.oidvci.CredentialIssuer
@@ -58,6 +60,7 @@ import io.ktor.http.*
 import io.ktor.util.*
 import kotlin.random.Random
 import kotlin.time.Clock
+import kotlin.time.Duration.Companion.minutes
 
 /**
  * Tests [OpenId4VciClient] against [CredentialIssuer] that uses [RemoteOAuth2AuthorizationServerAdapter]
@@ -238,15 +241,26 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
                 engine = mockEngine,
                 oauth2Client = OAuth2KtorClient(
                     engine = mockEngine,
-                    loadClientAttestationJwt = {
-                        BuildClientAttestationJwt(
-                            SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
-                            clientId = issuerPublicContext,
-                            issuer = "issuer",
-                            clientKey = issuerClientAuthKeyMaterial.jsonWebKey
-                        ).serialize()
+                    loadInstanceAttestation = {
+                        catching {
+                            BuildClientAttestationJwt(
+                                SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
+                                clientId = issuerPublicContext,
+                                issuer = "issuer",
+                                clientKey = issuerClientAuthKeyMaterial.jsonWebKey
+                            )
+                        }
                     },
-                    signClientAttestationPop = SignJwt(issuerClientAuthKeyMaterial, JwsHeaderNone()),
+                    loadInstanceAttestationPop = {
+                        catching {
+                            BuildClientAttestationPoPJwt(
+                                SignJwt(issuerClientAuthKeyMaterial, JwsHeaderNone()),
+                                clientId = issuerPublicContext,
+                                audience = issuerPublicContext,
+                                lifetime = 10.minutes,
+                            )
+                        }
+                    },
                     signDpop = SignJwt(issuerDpopKeyMaterial, JwsHeaderCertOrJwk()),
                     dpopAlgorithm = issuerDpopKeyMaterial.signatureAlgorithm.toJwsAlgorithm().getOrThrow(),
                     oAuth2Client = OAuth2Client(clientId = issuerPublicContext),
@@ -274,19 +288,42 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
                 engine = mockEngine,
                 oid4vciService = WalletService(
                     clientId = walletClientId,
-                    keyMaterial = credentialKeyMaterial,
+                    loadUnitAttestationPop = { input ->
+                        catching {
+                            SignJwt<JsonWebToken>(
+                                credentialKeyMaterial
+                            ) { header, material ->
+                                header.copy(jsonWebKey = material.jsonWebKey)
+                            }.invoke(
+                                input.type,
+                                input.payload,
+                                JsonWebToken.serializer(),
+                            ).getOrThrow()
+                        }
+                    }
                 ),
                 oauth2Client = OAuth2KtorClient(
                     engine = mockEngine,
-                    loadClientAttestationJwt = {
-                        BuildClientAttestationJwt(
-                            SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
-                            clientId = walletClientId,
-                            issuer = "issuer",
-                            clientKey = walletClientAuthKeyMaterial.jsonWebKey
-                        ).serialize()
+                    loadInstanceAttestation = {
+                        catching {
+                            BuildClientAttestationJwt(
+                                SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
+                                clientId = walletClientId,
+                                issuer = "issuer",
+                                clientKey = walletClientAuthKeyMaterial.jsonWebKey
+                            )
+                        }
                     },
-                    signClientAttestationPop = SignJwt(walletClientAuthKeyMaterial, JwsHeaderNone()),
+                    loadInstanceAttestationPop = {
+                        catching {
+                            BuildClientAttestationPoPJwt(
+                                SignJwt(walletClientAuthKeyMaterial, JwsHeaderNone()),
+                                clientId = walletClientId,
+                                audience = issuerPublicContext,
+                                lifetime = 10.minutes,
+                            )
+                        }
+                    },
                     signDpop = SignJwt(walletDpopKeyMaterial, JwsHeaderCertOrJwk()),
                     dpopAlgorithm = walletDpopKeyMaterial.signatureAlgorithm.toJwsAlgorithm().getOrThrow(),
                     oAuth2Client = OAuth2Client(clientId = walletClientId),